influx-db-sink/requirements.txt
git+https://github.com/quixio/quix-streams.git@feature/sinks
python-dotenv
influxdb3-python==0.3.6